Refactor project tracker page and stabilize hash-based row scrolling
