+++ app/project-tracker/page.tsx
@@ -2,13 +2,8 @@
 
 import Box from "@mui/material/Box";
 import Typography from "@mui/material/Typography";
-import type { TrackedProject, RepoStatus } from "@/app/lib/types";
-import {
-  transformGistData,
-  transformRepoData,
-} from "@/app/lib/githubTransformers";
-import { useEffect, useState } from "react";
-import { DataGrid, GridColDef } from "@mui/x-data-grid";
+import type { TrackedProject } from "@/app/lib/types";
+import { DataGrid } from "@mui/x-data-grid";
 import Paper from "@mui/material/Paper";
 import {
   Divider,
@@ -22,284 +17,38 @@ import {
 import StatusBadge from "@/app/component/status-badge";
 import { useCspNonce } from "@/app/component/csp-nonce-context";
 import NextLink from "next/link";
+import { projectTrackerColumns } from "@/app/project-tracker/component/columns";
+import { badgeDescriptions } from "@/app/project-tracker/lib/badges";
+import { useProjectTracker } from "@/app/project-tracker/lib/useProjectTracker";
 
-interface BadgeDescription {
-  id: number;
-  status: RepoStatus;
-  description: string;
+function ProjectTrackerNoRowsOverlay({
+  errorMessage,
+}: {
+  errorMessage: string | null;
+}) {
+  return (
+    <Box
+      sx={{
+        height: "100%",
+        display: "flex",
+        alignItems: "center",
+        justifyContent: "center",
+        px: 2,
+        textAlign: "center",
+      }}
+    >
+      <Typography color={errorMessage ? "error" : "text.secondary"}>
+        {errorMessage ?? "No rows"}
+      </Typography>
+    </Box>
+  );
 }
 
-type StrictProjectCol = Omit<GridColDef<TrackedProject>, "field"> & {
-  field: keyof TrackedProject;
-};
-
 export default function ProjectTracker() {
-  const nonce = useCspNonce();
+  const nonce: string | undefined = useCspNonce();
   const githubUsername: string = "StrangeRanger";
-  const [githubProjects, setGithubProjects] = useState<TrackedProject[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-
-  async function fetchAllRepos(): Promise<TrackedProject[]> {
-    const allRepos: Parameters<typeof transformRepoData>[0] = [];
-    let page: number = 1;
-    let hasMorePages: boolean = true;
-
-    while (hasMorePages) {
-      const response: Response = await fetch(
-        `https://api.github.com/users/${githubUsername}/repos?per_page=100&page=${page}`,
-        { cache: "no-store" },
-      );
-
-      if (!response.ok) {
-        throw new Error(`GitHub failed: ${response.status}`);
-      }
-
-      const data: unknown = await response.json();
-
-      if (!Array.isArray(data)) {
-        console.warn("Repos response is not an array:", data);
-        break;
-      }
-
-      // NOTE: `response.json()` is untyped. After confirming it's an array,
-      // we assert the element shape expected by `transformRepoData`.
-      // Use a runtime type guard if stronger validation is needed.
-      const repoPage = data as Parameters<typeof transformRepoData>[0];
-      allRepos.push(...repoPage);
-      hasMorePages = repoPage.length === 100;
-      page++;
-    }
-
-    return transformRepoData(allRepos);
-  }
-
-  async function fetchAllGists(): Promise<TrackedProject[]> {
-    const allGists: Parameters<typeof transformGistData>[0] = [];
-    let page: number = 1;
-    let hasMorePages: boolean = true;
-
-    while (hasMorePages) {
-      const response: Response = await fetch(
-        `https://api.github.com/users/${githubUsername}/gists?per_page=100&page=${page}`,
-        { cache: "no-store" },
-      );
-
-      if (!response.ok) {
-        throw new Error(`GitHub failed: ${response.status}`);
-      }
-
-      const data: unknown = await response.json();
-
-      if (!Array.isArray(data)) {
-        console.warn("Gists response is not an array:", data);
-        break;
-      }
-
-      // NOTE: `response.json()` is untyped. After confirming it's an array,
-      // we assert the element shape expected by `transformGistData`.
-      // Use a runtime type guard if stronger validation is needed.
-      const gistPage = data as Parameters<typeof transformGistData>[0];
-      allGists.push(...gistPage);
-      hasMorePages = gistPage.length === 100;
-      page++;
-    }
-
-    return transformGistData(allGists);
-  }
-
-  function ProjectTrackerNoRowsOverlay({
-    errorMessage,
-  }: {
-    errorMessage: string | null;
-  }) {
-    return (
-      <Box
-        sx={{
-          height: "100%",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-          px: 2,
-          textAlign: "center",
-        }}
-      >
-        <Typography color={errorMessage ? "error" : "text.secondary"}>
-          {errorMessage ?? "No rows"}
-        </Typography>
-      </Box>
-    );
-  }
-
-  function handleHashScroll() {
-    const hash: string = window.location.hash;
-
-    if (!hash) return;
-
-    const rowId: string = decodeURIComponent(hash.slice(1));
-
-    // Wait a bit to ensure the browser has fully rendered the page.
-    setTimeout(() => {
-      const selector = `[data-id="${CSS.escape(rowId)}"]`;
-      const element = document.querySelector(selector);
-      if (element) {
-        element.scrollIntoView({
-          behavior: "smooth",
-          block: "center",
-        });
-      }
-    }, 100);
-  }
-
-  useEffect(() => {
-    let isCancelled: boolean = false;
-
-    async function run(): Promise<void> {
-      try {
-        const [repos, gists] = await Promise.all([
-          fetchAllRepos(),
-          fetchAllGists(),
-        ]);
-        if (!isCancelled) {
-          setGithubProjects([...repos, ...gists]);
-          setErrorMessage(null);
-        }
-      } catch (error) {
-        if (!isCancelled) {
-          const message: string =
-            error instanceof Error ? error.message : "Unknown error";
-          setErrorMessage(message);
-        }
-      } finally {
-        if (!isCancelled) {
-          setIsLoading(false);
-          handleHashScroll();
-        }
-      }
-    }
-
-    run();
-    return () => {
-      isCancelled = true;
-    };
-  }, []);
-
-  const columns: StrictProjectCol[] = [
-    { field: "name", headerName: "Project Name", flex: 1, minWidth: 200 },
-    { field: "projectType", headerName: "Type", width: 120 },
-    {
-      field: "status",
-      headerName: "Status",
-      width: 150,
-      renderCell: (params) => {
-        // NOTE: DataGrid cell values are broadly typed; this cast narrows the
-        // `status` cell value to our `RepoStatus` union for `StatusBadge`.
-        // Prefer typed render params if we want to avoid assertions entirely.
-        const status = params.value as RepoStatus;
-        return <StatusBadge status={status} />;
-      },
-    },
-    {
-      field: "lastCommitTimestamp",
-      headerName: "Last Commit",
-      width: 180,
-      renderCell: (params) => params.row.lastCommitRelative,
-    },
-    { field: "description", headerName: "Description", flex: 2, minWidth: 300 },
-  ];
-
-  const badgeDescriptions: BadgeDescription[] = [
-    {
-      id: 1,
-      status: "personal",
-      description:
-        "Indicates that the project or document is tailored to my personal needs and will be updated as required. Typically, it reflects my current setup, usage, or preferences.",
-    },
-    {
-      id: 2,
-      status: "active",
-      description:
-        "The project is stable and fully functional, with ongoing development and regular updates. New features and improvements are continuously being added.",
-    },
-    {
-      id: 3,
-      status: "maintained",
-      description:
-        "The project is stable and functional, receiving updates primarily for bug fixes and minor improvements. Active development is minimal but ongoing as needed.",
-    },
-    {
-      id: 4,
-      status: "inactive",
-      description:
-        "Development on the project has paused, but it remains in a stable and usable state. Future work may resume, but there are currently no active updates or enhancements.",
-    },
-    {
-      id: 5,
-      status: "finished",
-      description:
-        "The project is complete and fully functional. While no active development is planned, updates may occur if essential fixes or changes are necessary. Combines aspects of both Maintained and Unsupported statuses.",
-    },
-    {
-      id: 6,
-      status: "unsupported",
-      description:
-        "The project is stable and usable, but active development has ceased. No further updates are planned, and users may need to seek alternative maintainers or solutions if issues arise.",
-    },
-    // NOTE: Disabled for now, though it may be reintroduced later.
-    // {
-    //   id: 7,
-    //   status: "continuous",
-    //   description: "The project is under ongoing development with a focus on gradual improvements and enhancements. The development pace is steady but less rapid than that of active projects, blending elements of active and maintained statuses.",
-    // },
-    {
-      id: 7,
-      status: "concept",
-      description:
-        "Represents an early-stage project or proof-of-concept with minimal implementation. Intended for demonstration, experimentation, or initial exploration without full functionality.",
-    },
-    {
-      id: 8,
-      status: "wip",
-      description:
-        "Development is actively underway, but the project has yet to reach a stable or publicly usable state. Ongoing work is focused on achieving initial functionality and stability.",
-    },
-    {
-      id: 9,
-      status: "suspended",
-      description:
-        "Development has been temporarily halted after initial progress. The project remains in a usable state, with intentions to resume work in the future, pending circumstances.",
-    },
-    {
-      id: 10,
-      status: "abandoned",
-      description:
-        "The project has been discontinued and will no longer receive updates or support. Users are encouraged to seek alternatives or fork the project if continued development is desired. Assume the project has been archived.",
-    },
-    {
-      id: 11,
-      status: "archived",
-      description:
-        "The project has been officially archived, meaning it is no longer maintained or supported. It serves as a historical reference, and no further changes will be made unless specified otherwise.",
-    },
-    {
-      id: 12,
-      status: "moved",
-      description:
-        "The project has been relocated to a new repository or platform. The new location is the authoritative source, and all future updates and maintenance will occur there. Assume the project has been archived.",
-    },
-    {
-      id: 13,
-      status: "unspecified",
-      description:
-        "The project status is not explicitly defined or documented. This status is usually reserved for projects that don't need a specific status. For example, a <a href='https://github.com/StrangeRanger/StrangeRanger' target='_blank'>GitHub user's public profile page</a> will have this status, as it's not a traditional software project.",
-    },
-    {
-      id: 14,
-      status: "unknown",
-      description:
-        "The project status is not known or has not been determined. This status is typically used when the project is new, lacks documentation, is a fork, or has not been categorized yet.",
-    },
-  ];
+  const { githubProjects, isLoading, errorMessage } =
+    useProjectTracker(githubUsername);
 
   return (
     <Box>
@@ -318,7 +67,7 @@ export default function ProjectTracker() {
         </Typography>
         <Paper sx={{ width: "100%" }}>
           <DataGrid
-            columns={columns}
+            columns={projectTrackerColumns}
             rows={githubProjects}
             getRowId={(row: TrackedProject) => `project-${row.id}`}
             autoHeight
